refactor some util functions
