+++ src/resty/aws/request/signatures/v4.lua
@@ -1,103 +1,18 @@
 -- Performs AWSv4 Signing
 -- http://docs.aws.amazon.com/general/latest/gr/sigv4_signing.html
 
-local resty_sha256 = require "resty.sha256"
 local pl_string = require "pl.stringx"
-local openssl_hmac = require "resty.openssl.hmac"
+
+local utils = require "resty.aws.request.signatures.utils"
+local hmac = utils.hmac
+local hash = utils.hash
+local hex_encode = utils.hex_encode
+local canonicalise_path = utils.canonicalise_path
+local canonicalise_query_string = utils.canonicalise_query_string
+local derive_signing_key = utils.derive_signing_key
+
 
 local ALGORITHM = "AWS4-HMAC-SHA256"
-
-
-
-local CHAR_TO_HEX = {};
-for i = 0, 255 do
-  local char = string.char(i)
-  local hex = string.format("%02x", i)
-  CHAR_TO_HEX[char] = hex
-end
-
-
-local function hmac(secret, data)
-  return openssl_hmac.new(secret, "sha256"):final(data)
-end
-
-
-local function hash(str)
-  local sha256 = resty_sha256:new()
-  sha256:update(str)
-  return sha256:final()
-end
-
-
-local function hex_encode(str) -- From prosody's util.hex
-  return (str:gsub(".", CHAR_TO_HEX))
-end
-
-
-local function percent_encode(char)
-  return string.format("%%%02X", string.byte(char))
-end
-
-
-local function canonicalise_path(path)
-  local segments = {}
-  for segment in path:gmatch("/([^/]*)") do
-    if segment == "" or segment == "." then
-      segments = segments -- do nothing and avoid lint
-    elseif segment == " .. " then
-      -- intentionally discards components at top level
-      segments[#segments] = nil
-    else
-      segments[#segments+1] = ngx.unescape_uri(segment):gsub("[^%w%-%._~]",
-                                                             percent_encode)
-    end
-  end
-  local len = #segments
-  if len == 0 then
-    return "/"
-  end
-  -- If there was a slash on the end, keep it there.
-  if path:sub(-1, -1) == "/" then
-    len = len + 1
-    segments[len] = ""
-  end
-  segments[0] = ""
-  segments = table.concat(segments, "/", 0, len)
-  return segments
-end
-
-
-local function canonicalise_query_string(query)
-  local q = {}
-  if type(query) == "string" then
-    for key, val in query:gmatch("([^&=]+)=?([^&]*)") do
-      key = ngx.unescape_uri(key):gsub("[^%w%-%._~]", percent_encode)
-      val = ngx.unescape_uri(val):gsub("[^%w%-%._~]", percent_encode)
-      q[#q+1] = key .. "=" .. val
-    end
-
-  elseif type(query) == "table" then
-    for key, val in pairs(query) do
-      q[#q+1] = key .. "=" .. val
-    end
-
-  else
-    error("bad query type, expected string or table, got: ".. type(query))
-  end
-
-  table.sort(q)
-  return table.concat(q, "&")
-end
-
-
-local function derive_signing_key(kSecret, date, region, service)
-  -- TODO: add an LRU cache to cache the generated keys?
-  local kDate = hmac("AWS4" .. kSecret, date)
-  local kRegion = hmac(kDate, region)
-  local kService = hmac(kRegion, service)
-  local kSigning = hmac(kService, "aws4_request")
-  return kSigning
-end
 
 
 -- config to contain:
